feat: Refactor nutrition card to be splitted into cards
