+++ lib/domain/models/enums/nutrition_card_types.dart
@@ -1,0 +1,7 @@
+enum NutritionCardType {
+  mealSelector, // For breakfast/lunch/dinner with dropdown
+  foodItem, // Individual food items with remove option
+  addButton, // Add new item button
+  inputField, // For manual nutrition input
+  searchResult, // New type for search results like in the image
+}
